fix(feature/mypage): diary로 워딩 변경, 무한 스크롤 문제 수정
